fix: 함수나 파일구조 정리
