+++ src/constants/index.ts
@@ -1,1 +1,2 @@
 export const OPERATION_REGEX = /([/+\-X=])/g;
+export const CALCULATOR_NUMBERS = [...Array(10).keys()];
